10.06

+++ src/main/java/com/example/app/service/SlicerServiceImp.java
@@ -2,13 +2,15 @@ package com.example.app.service;
 
 import com.example.app.db.entity.*;
 import com.example.app.db.repository.SlicerRepository;
-import com.example.app.untils.Generator;
-import org.apache.catalina.User;
+import com.example.app.until.Generator;
+import org.apache.logging.log4j.CloseableThreadContext;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
+import java.time.Instant;
+import java.time.LocalDate;
 import java.util.List;
 
 @Service
@@ -22,7 +24,7 @@ public class SlicerServiceImp implements SlicerService {
     public String slice(String url, String username) {
 
         UserData userData = userService.findUserByName(username);
-        String slicedUrl = Generator.generateId();
+        String slicedUrl = generateUniqueCode();
 
         UrlData newUrl = new UrlData(slicedUrl, url, userData);
         repo.save(newUrl);
@@ -31,9 +33,16 @@ public class SlicerServiceImp implements SlicerService {
     }
 
     @Override
-    public UrlData edit(String slicedUrl, String newTrueUrl) {
+    public UrlData edit(String slicedUrl, String newTrueUrl, String user) {
         UrlData url = repo.findById(slicedUrl)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Link not found"));
+
+        if(!url.getUsername()
+                .getUsername()
+                .equals(user)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You can not do that to this link");
+        }
+
         url.updateUrl(newTrueUrl);
         return repo.save(url);
     }
@@ -51,9 +60,16 @@ public class SlicerServiceImp implements SlicerService {
     }
 
     @Override
-    public UrlData delete(String id) {
+    public UrlData delete(String id, String user) {
         UrlData deletedData = repo.findById(id)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Link not found"));
+
+        if(!deletedData.getUsername()
+                .getUsername()
+                .equals(user)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You can not do that to this link");
+        }
+
         repo.deleteById(id);
         return deletedData;
     }
@@ -66,8 +82,22 @@ public class SlicerServiceImp implements SlicerService {
         if(!url.getIsActive()){
             throw new ResponseStatusException(HttpStatus.LOCKED, "This link is not active");
         }
+
+        if(LocalDate.now()
+                .isAfter(url.getDate())) {
+            throw new ResponseStatusException(HttpStatus.GONE, "Link is expired");
+        }
+
         repo.incrementRefCount(id);
 
         return url.getUrl();
+    }
+
+    private String generateUniqueCode() {
+        String code;
+        do {
+            code = Generator.generateId();
+        } while (repo.existsById(code));
+        return code;
     }
 }
